re-usable fixture helper methods
